pyproject.toml
[project]
name = "wanchain-iwan"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "logger>=1.4",
    "pandas>=2.2.3",
    "pangres>=4.2.1",
    "pyyaml>=6.0.2",
    "sqlalchemy>=2.0.38",
    "websockets>=15.0.1",
    "psycopg[binary]>=3.2.5",
]
